Fix bug #181081: "avatar is caught in an unreachable area"
When Kapman makes a half-turn right after taking a turn at a corner, it may
go through an opposing wall.

This fix is rather ugly, but a better solution would probably need quite
some refactoring. So, just fix it for now.

Also merge duplicate code paths in the same function.

BUG: 181081